12/04/2020 Adding Button Functionality and Alerts
Summary
Adds functionality to some of the buttons on the home screen and tracker screen. This will allow the user to enter medicine of their own as well as being able to view information on the included drugs. Beginning the functionality of signing up a new user and verification of whether a user or drug entered is valid

Query Params:
Query Params has been used to achieve the extra information screen which imports information from the Medicine screen. This is done by transferring a special variable and referncing this in the information. This was achieved by creating the following function to execute in tabs2.page.ts
onInfoPress(index, med, medicines, descriptions) {
    index = this.medicines.indexOf(med)
    console.log(medicines[index])
    console.log(descriptions[index])
    console.log(descriptions);
    let navigationExtras: NavigationExtras = {
      queryParams: {
        special: med.desc

      }
    };
    this.router.navigate(['details'], navigationExtras);
and the following in tab2.page.html, note the parameters which in future development I hope to decrease the reliance/need for them:
<ion-button size="large" (click)="onInfoPress(index, med, medicines,descriptions)">
          <ion-icon name="information-circle-outline"></ion-icon>
</ion-button>

Alerts:
Alerts have now been included. This is to "alert" the user that the information they have entered either in the login boxes or the medicine section.

For the login screen it determines whether the user information entered is valid and if not signals an alert to be displayed to let the user that they need to enter valid details. This is achieved by:
#Checking whether a user's login is valid
 for (let i of this.userLogin) {
      if (i.username == this.inputtedUsername) {
        if (i.password == this.inputtedPassword) {
          console.log("Gained entry")
        }
        else {
          this.alertPresent();
        }
      }

      else {
        this.alertPresent();
      }
    }

#this.alertPresent() activated the alert with the error message
async alertPresent() {
    const alert = await this.alertController.create({
      header: "Alert",
      subHeader: "Incorrect Username/Password",
      message: "Please make sure that you enter a valid username and password combination",
      buttons: ['Dismiss']
    });

    await alert.present();
    let result = await alert.onDidDismiss();
    console.log(result);

To do this the 2 input boxes have been bound using ngModel to “inputtedUsername” and “inputtedPassword”. This will allow the elements to be checked and look if the values are valid or not.
<h2>Username</h2>
      <ion-item>
        <ion-input type="text" placeholder="Enter your Username" [(ngModel)]="inputtedUsername">
        </ion-input>
      </ion-item>
      <h2>Password</h2>
      <ion-item>
        <ion-input type="password" placeholder="Enter your password" [(ngModel)]="inputtedPassword">

        </ion-input>
      </ion-item>

User Sign Up:
This uses an alert as a sign up sheet for a user. The indicated fields have not been modified to truly reflect what will be entered and will be updated in the future. To ensure duplicate users are not created. As the “ok” button is pressed the current list of valid usernames and passwords is checked to ensure that there is no existing user before allowing a profile to be pushed to the saved array of userLogins. Note that the userLogins array has been updated to include [“username”] and [password”] attributes, this will greatly help verification checking. To help usability requirements the background has been set so that it does not close the alert down if clicked, as this would be very annoying halfway through signing up!
#Sign Up Button Functionality and Alert
async presentSignUp() {
    const alert = await this.alertController.create({
      backdropDismiss: false,
      header: 'Sign Up',
      inputs: [
        {
          name: 'uName',
          type: 'text',
          placeholder: 'Placeholder 1'
        },
        {
          name: 'pWord',
          type: 'text',
          id: 'name2-id',
          value: 'hello',
          placeholder: 'Placeholder2'
        },
        {
          name: 'name3',
          type: 'date',
          min: '2017-03-01',
          max: '2018-01-12'
        }
      ],
      buttons: [{

        text: 'Cancel',
        role: 'cancel',
        handler: () => {
          console.log('Confirm Cancel');
          this.determine = 'Cancel';
        }
      },
      {
        text: "Ok",
        role: 'Ok',
        handler: () => {
          console.log('Confirm Ok');
          this.determine = 'Ok';
        }
      }
      ],

    });
    await alert.present();
    let result = await alert.onDidDismiss();
    console.log(result);
    console.log("This is the role: " + result.data.values.uName);
    console.log("This is the first userLogins ", this.userLogin);
    if (this.determine == "Ok") {
      let prof: any = new Object();
      prof["username"] = result.data.values.uName;
      prof["password"] = result.data.values.pWord;
      this.userLogin.push(prof);
      console.log("This is the updated userLogins ", this.userLogin);
    }
Similar to the user sign up the drugs sign up is an alert
Variables created intab2:
addmedicine="";
 adddescription="";
 savedMedicine=[];
 add=true;
 determine: string;

These are to help addmedicine and their relevant descriptions to the existing array as well as having an array set up ready for the storage of “saved Medicines” (ones which the user presses the addition button on). Add is a variable which changes value depending on whether a medicine can be added or not and if a medicine is a duplicate then it is set to false and cannot be edited
The changes on Tab 3 are purely experimental and have no relevance to the current project at this current point in time.
SCSS:
Tab 1:
Padding-left and right have been left unchanged for the buttons for now aside from an attempted change in font-size. Formatting of the user input boxes remains central
Tab 2:
The information button and medicine divs have been set that they give the app a slight border and the app does not go upto right to the edge of the screen