14/04/2020 Toggle Switch and localStorage
Tab1.page.html
This has been modified to interact with localStorage. At this point in time it uses functions called saveItem() and retrieveItem() with test data to ensure that the functions are set up to handle storage of objects. This has been altered to display the saved list to see if the functionality of these buttons works. This is temporary and will later be removed and placed on the “Tracker” page (tab3).
Tab1.page.ts
As mentioned above saveItem() and retrieveItem() have been added which now successfully store arrays of information and retrieve them in the correct format. This will allow saved lists of drugs which are related to a specific user to be displayed on the screen.
A display profile function has been created which closes down the html div with the login screen information and opens up a profile screen. I have since decided that the landing page will switch over to an “Information” screen on how to use the application and a separate Profile tab. This is currently in development.
Tab2.page.html
In this update I have switched the addition feature from a button to a toggle switch. This is still going to have the same functionality as the button which is to add a medicine to their saved list. This has been changed as I deemed a toggle switch to be more usable as it is clear what state each drug is in (whether is has been added or not to the list) compared to an average button which it can easily be forgotten whether a user has added a certain medication to a list.
Tab2.page.ts
This led to the development of the testing function. This checks what state the toggleswitch is in and carries out the appropriate action. If the switch is on the drug is added to the saved list, if it is off it is removed from the saved list. This is done by each drug having a toggle switch variable which is amended depending on the state of the switch.
An alert for signing up a new drug was created. This has entry fields for the Drug Name and Description. Depending on whether “Ok” or “Cancel” is pressed the drug is either added to the list to display on the page (If “Ok” pressed) or the information discarded (If “Cancel” pressed). BackdropDismiss will be enabled to ensure that it is impossible for a user to accidentally click off their entry halfway through filling it in. This will increase the usability.
Tab3.page.html
This has been altered to display a button to update a saved list array and later display it onto the screen. Note that there is a “Retrieve button” function on this page. This is so that the saved list from the other pages (tab1 and tab2) can be loaded into Tab3 and displayed accordingly).
Tab3.page.ts
Reflects this functionality and updates the saved array accordingly before being displayed. At the moment the data retrieved from the retrieveItem() function is not properly used but this will be integrated by the next update.
Tabs.page.html
I have changed the names of the tabs so that they reflect what is contained on each tab. This will help usability as user’s will find it easier to understand which tab entails.