15/04/2020
Tab1.page.ts
Comments have been added to all of the functions to explain what they do. This has helped improved code maintenance and allow future developers to understand what each function does. The code has also been cleared up by removing the variables and arrays that were no longer needed in the functioning of the application.
Further improvements have been made in regards to using and storing data to the local storage where it is appropriate. This has led to the rest of the pages following the same protocol meaning that all access and change the same profile to help ensure consistency with the same user profile across the application.
Tab2.page.html
In this the onInfoPress command has been updated to delete some of the parameters that are no longer needed
Tab2.page.ts
A clean up of all of the variables has occurred. Getting rid of the ones that were not necessary for the functioning of the application.
All of the functions use the local storage now so have been adapted so that they retrieve the stored currentUser array from the local storage and save the changes that they make to it afterwards (such as adding a new medicine or tracking a new one. This was the main objective completed with this update.
Tab3.page.html
This has been updated to include a button that updates the tracked array list and displays it on the screen this information is retrieved from local storage. For each component I have added a ion-card as I believe this will increase the usability of the application as it systematically splits up each object into it’s separate card which helps the user see which information belongs to what medicine.
Tab3.page.ts
The updateItems() and retrieveItems() functions have been added. This will ensure that the currentUser profile can be loaded and the relevant information (in accordance to the currentUser) is displayed on their tracked medicine screen