16/04/2020
Login Page:
The code located within this page’s files is an exact duplicate of the previous versions of the tab1 files. This has been moved over to a dedicated login page. This will be called upon using the modal controller to display the log in screen whenever a user is not logged into the application.
App.Module.ts
The LoginPageModule has been imported so it is accessible for routing
App-Routing-Module.ts
This has added the child of the login page. This is so that it can be controlled and used by the modal controller to display the login screen when the app is run.
App.Component.ts:
There have been significant changes in this file in order for the modal controller to work effectively. This works by retruieved the current user and seeing if there is a user present and if there is the login modal need not be displayed. However if there is no user or an invalid user present the login screen is presented.
As previously mentioned this has been achieved by using a modal controller which creates a component and does the relevant checks to see whether the login portal needs to be displayed. The backdrop has also been set to dismiss so a logged out user can’t just exit the login page and break the app.
Tab1.page.html
This has been transformed into a help page. This is to be a separate tab that the user can choose to seek guidance on how to function the application. For example: This gives a visual prompt of a button and explains what the functionality of it and where it passes the information to and displays the medication. This will increase usability as users will know where they need to go and get help from as it will be clear
Tab1.page.ts
This will be required to be deleted and will be dealt with in a futher update as none of the ts has an impact on the html help screen any longer
Tab2.page.html
The click function name for the toggle switches was changed from testing to something more appropriate (toggled). The functionality has not changed just the name of the function.
A searchbar has also been added, which is not a required but would be desirably feature as this would increase usability.
Tab4.page.html
A sign out button has been added which will allow a user to save their data and exit the application. The title for the page has been changed also to Profile to reflect what will be contained on this page.
Tab4.page.ts
The signOut() function has been created which will allow the user’s data to be saved and the currentUser() reset to null so that the login screen is displayed again.
Tabs.page.html
The name and the icons of all of the pages have been altered. This is to ensure that user’s will find it easier to recognise what the contents of individual pages will contain. In turn increasing the usability of the application.