20/04/2020
App.module.ts
The HttpClientModule and DataService functions have been important and the DataService has been listed as a provider in preparation for use within the application
Data.service.ts
This allows us to user the http client to create a function that will allow us to create a query string for our api to use. This has been called APIGetSynonyms
Details.page.html
This has got an input box, ready for the user to enter an unknown word that they want to look up using the API. This requires a “Search” button to be clicked which will activate searching the querystring using the api that has been imported. This currently displays a lot of information and in future this will be changed so that a limited number of results appears
Word-chain.ts
This has been created in order to help with the functionality of the API created. This includes listing the elements that are involved in a single data item
Tab4.page.ts
This has been updated to include a functioning signOut options which deletes a user’s data upon them logging out. This is carried out after the currentUser’s settings and profile (medication and timer settings have been saved to a userLogin profile stored in the local storage. This is vital for the login/signout feature to work effectively.
A local notification system has been attempted in this update. This requires finetuning for it to work properly and uses the local notification import. This is being tested so that it can be implemented for the tracker screen and the countdown timers present there.
Tab3.page.ts
This now has working timer functionality which starts and stops depending on a user’s entered time of the day which they want to set an alert for. Note at this moment in time the timer is a 24 hour daily one and there is no other setting that is currently active. It is also worth noting that at this current point in time the countdown timer only effectively works for one set time which is then shared by all of the other timers. This will be patched in the future.
The datetime function accurately calculates the amount of time that there are between the user’s inputted time and the current time meaning the timer can be started and countdown from this amount and display this time on the screen
Tab3.page.html
This has been updated to display the timer and the relevant drug name that it should come under to make it easy to display and allow the user to see which drug is assigned which timer.
Tab2.page.ts
This has been updated to retrieve items and set arrays depending on the currentUser. A remove medicine button has also been added in case the user accidentally adds a medicine and they want to remove it from their version of the app. This removes the desired medicine from the userAccount and saves the information to local storage.
Tab2.page.html
A remove button has been added and given an appropriate name and colour for the user to clearly see what this button is to carry out.
Tab1.page.ts
This has received a new alert which is to be displayed when the login that a user has attempted to carry out fails.
NavController/Route was imported as there was a problem with the re-direction from getting to/from the login screen.
Tab1.page.html
This has been updated to provide a “get logged in” in case the user is not automatically directed to the login screen on opening the application.
login.page.ts
This has been updated to provide an update to how new user’s are created and what information they are signed. Aside from this minor changes are made to variables and arrays that are used.
A delete user data button has been added for testing purposes and this clears all user accounts from the userLogins array
App.routing.module
Paths for the different tabs have been re-configured to ensure that they load the correct tab when it is clicked on.

In the next update it is hopeful that this project will be integrated across to a sidemenu app. At this current point in time the app uses HTML storage and not the Ionic Storage meaning that it is not efficient and does not effectively retain or update userLogins as required for the application to work. Alongside this the modal controller seems to be very buggy and so instead of implementing this I will be using the code (updating and cleaning as I go) from this project to create a sidemenu app which uses all Ionic storage capabilities. This will be seen starting from the next update.