21/04/2020
This is set up using a side-menu interface as whilst I was previously developing this application there were elements that just felt they could be achieved by using a more efficient layout, not requiring tabs at the bottom. This gives more space for information to be displayed and less risk of a tab that a user does not want to be on being clicked.
The relevant pages are being created such as Login/Profile, Medication, Tracker and Help page. At this point this is not an exhaustive list of pages that might be created and this may be increased at a later date to accommodate new pages.
In this update the login page has been transferred across to the new interface. This allows the user to create and register as previously done before storing their data in local storage. The key difference now being that the local storage used is through Ionic. This allows the storage to be cleaner and allow easier data retrieval/saving capabilities once the code is turned into the apk.
Divs have been set up to display a login screen if the user is not logged in and if they are logged in then the Profile screen is displayed. This makes the application usable as the user will not have to try and dismiss the login screen everytime they open the app unless they have logged themselves out.