24/04/2020
At this current point in time the functionality that has been added has been transferred across from the old template. It is in the process of being optimised for this template.

The Details page has been complete in an identical way, using query params to transfer data across to the relevant drug to receive a description of what it is.

The medication page has been altered so that now it has the functionality to add new medicine as well as delete them. Verification and validation occurs when a user enters a “new” medicine to check whether it already exists and denies the user the ability to add it if it does. This increases the usability because multiple iterations of the same drug can’t appear on the list. This is a text based method so if there are 2 different names for the same drug it is worth noting that the drug will still be added. The addition button works and adds the drug to the tracker page, this successfully syncs across and has been test by added code to the tracker ts and html to retrieve and display this relevant list