29/04/2020
I have created a database to be accessed in local storage called “_myDB”, this is in order for data to be stored more efficiently and it is known where data is to be stored.
I have created global variables which can be accessed throughout the application. This will limit the need for constant saving and retrieving of information (which can take extra time to do and slow the working of the application down).
The help page has been updated to show information to the user on how to use this application. At this point in time there is no dynamic capability to display a certain screen depending on if there is a user is logged in or not but this will be added.
There was a minor change to the login page to test whether retrieving the user was being done effectively from the ionic storage. It was concluded that it was not and that global variables should be used more frequently in the development.
The login page ts has been updated to ensure that the global variables are imported on the page being loaded and these global variables being updated throughout the code before the currentUser and the userLogins array being updated with the latest settings. This is done on log out and the usability of the application might be improved if there is a “Save settings” button to ensure nothing is lost upon a force close or similar of the application.
The medication page has been altered to present an alert if the user is not logged in to redirect them to the login screen. This can’t be dismissed and the user is only given one option, to go to the login screen. This will help usability as a user will then know they are required to complete an action on the login page.
The tracker page ts and html have been updated so that the timer feature has functionality and works effectively. This is using similar code to that of the tabbed interface application.