03/05/2020
This update finalises the application to work properly. The first thing to note is that the resources have been updated and that a unique icon and splashscreen have been added to get the mood for the application started as it is loading.
For all of the help pages routing has been added to the app-routing-module in order for these pages to be accessed via links. This has been done by creating relevant paths which load the correct “child” page.
All of the pages have been updated with relevant scss and styles within the html documents to ensure that they are consistent across the application and alternate accordingly.
The following packages have been imported so that they can be used across the application:
•	Camera
•	LocalNotifications
This is carried out within the app.module.ts and allows these packages to be used to increase the usability of the application. Without the local notifications imported the application will not be allowed to send push notifications to the user to remind them to take their medication.
In the details page the content has been updated to ensure that the correct images are shown on the page. A camera icon has also been added to ensure that the user is aware that they can add a picture. This will be useful for usability. The typescript has been updated accordingly to use the camera package and update the global photos array to display the photos that are taken depending on the drug that it has been taken for.
For each of the following pages:
•	Details
•	Profile
•	Login
•	Medication
•	Tracker
A help screen has been created which has tips on how to use each page retrospectively, this will allow a user to check what a feature may do if they are unsure how to use the app.
More global constants have been created to allow access from multiple typescripts. A class for photos has been created, this is in order store and convey photographs for the drugs.
The help page has been equipped so that it dynamically loads a help page. If you are logged out it shows a help page on how to get logged on and if the user logs in then the rest of the information for all the pages is displayed and not the login page help. They have retrospective route links to the relevant pages.
The login page has been updated to update the look and provide a cleaner and more usable look for the application. The colour scheme for the application has been modified and the position and size of the buttons and information has been changed also to provide a better experience
For the medication page it has been updated to contain ion-cards which separate the distinct information to increase usability, to make it clear which buttons relate to actions on which medication. The typescript has also been updated to provide feedback on what a particular action has consequences on.
The tracker page has been updated to dynamically load the tracked medicine if it is there and display a help message if a medication is not added due to how a user may not know how to add it as it would have to be added on the medication screen and not the tracker screen. The remove tracker medicine function has been added which allows a user to remove the medicine which is currently located on the page